tema3/image-processing-app/cloud-function/requirements.txt
google-cloud-vision
google-cloud-firestore
gunicorn